added new test_message
Ignore-this: 9ad335091626fd869e291767307a3dee

This is a test to see how to write a chat program and which pieces are nessasary.
Added a sender object which sends messages appending the user id metadata.

This is a step to creating the idea of a session and having an app library

Also removed firetest because now firestr can send messages

darcs-hash:20120831021358-c21fe-25bcb414e7d2367b2e99c3fe479efcd8e9f159db