replaced how contact status is updated
Ignore-this: aa69ed9b139a2f2c502008274f5d06df

I switched to useing events to update the contact lists in sessions instead of
the QTimers per list. This results in immediate feedback as soon as user_service
knows the status of a contact.

I also moved the primary app event loop from messagelist to session because it didn't
make sense for messagelist to control the event loop for a session.

I fixed a bug in user_service which failed to update the status to disonnect when
a user pinged a disconnect message.

Now when a user disconnects on purpose, everyone knows about that immediately.

darcs-hash:20120914190250-c21fe-ef60095ca3e5d7bf592ff5adfd52a975e6995c31