added preliminary connection manager
Ignore-this: f7dc2cd849be5b0a8883bc6f3d76a14c

At this stage I got a listen socket and outgoing connections working on the same port.
I added a preliminary connection manager which currently has a pool of bound sockets
which can be used to make out going connections from the same port the listen socket
is on.

The next step is for the connection manager to keep track of which outgoing or incoming
connections correspond to a working socket with a peer and use the appropriate one
to communicate over the network.

This means any given connection with a peer will be a client or server connection and
any firestr instance will act as both in any given moment.

The trick with the connection manager is that the bound pool is created before the
listen socket is created. This is because certain TCP implementations won't let you
bind to a port with an already listening socket... So if we bind before we listen
things work...

darcs-hash:20121011023625-c21fe-507c03fd3044f6d927dc4129043620f2629c2ce4