more progress towards sending messages over udp
Ignore-this: bf41a3e2c33eb489ec708e7c48484296

The code is in a compilable state and some of the basic data structures are there.
I still need to implement the mechanics of sending and reading a message.

The idea is that many messages might not fit withing a UDP packet so messages are
sent in chunks. And because udp does not gaurentee delivery or order, I will
have to keep track of that on the reading end. and have some sort of timeout
for messages that don't get all their chunks.

I am not sure if I want to implement a resend functionality like tcp...
Maybe I can gaurentee order of delivery, but not actual delivery.
Still thinking about it...

darcs-hash:20121120142348-c21fe-21acda5b38a4113eb8d71917655b046d23f18375