added ability for apps to interact with each other in a session
Ignore-this: 3b54d8e39546042212df45de253504df

You can now send messages locally in a session to other apps. There is a way
to iterate through all apps in a session and send them messages. There is also
a new send_local function to send a message to all apps in a session.

Now we have a way to communicate apps across the network and between each other.
This should be really powerful if apps are written to work together.

darcs-hash:20130401202524-c21fe-df6d0092c24fe942024146a1e43a766997eaae52