added reconnect every minute
Ignore-this: b8a9f97e07fddb437d7dd21959664b41

This may be too short a time. But this is for the case where network connection is lost,
or computer goes to sleep. We want to reconnect instead of having to quit the app and restart
it.

I created a new thread which reconnects every minute by initiating the startup procedure
again that was previously done.

Only peers that are not connected are sent greet messages...

darcs-hash:20131018192635-c21fe-5baab5cecabbc84c4e029790c90475d998149253