removed disconnect message from session
Ignore-this: e277aa0cf03306c7bb7b50c67738365d

The UI now does not announce when a user disconnects in a session tab. Because we now
have auto reconnect, you can leave a session tab open for days and still successfully reconnect
to a peer and have the session applications communication restored. Since the user
turns red when they disconnect and since it is already announced in the alert tab, I
thought this message was silly.

The message when a user explicitly quits a session is still there.

darcs-hash:20131106180353-c21fe-e45af2bad872afa645362339de43919879b8789f