added a status widget to show state of app editor
Ignore-this: 18be01fa294b60b52a5630a2c89e1b17

This is so that the user can understand when the editor is waiting for typing
to finish and when code is compiled. Sometimes you can write code and the UI
for the app does not change, but the user had no way to know if the code was
compiled. Now with the status widget, there is some feedback.

darcs-hash:20131202151545-c21fe-dc639bedbda6130b596239596c926d9cf047116f