fixed alerts for windows out of focus
Ignore-this: c82e003dee612f24cc9577558b6baf35

Now if the main window is out of focus, an alert turn the tab red even if it is
the currently selected tab. Also won't turn a tab red if the window has focus and
the current tab selected is the same as the alert.

When you focus a window, the alert is now properly turned off by calling tab_changed on focus.

darcs-hash:20140122192710-c21fe-f4f993c7eb29b75807122126a8a9c233a131cf22