only can add contact to a conversation if they are part of a clique
Ignore-this: be7bb6aabb102c499fe09a7835b3fe99

This is a step to allow guests in a conversation. But first we need a way to determine
if a contact is part of a clique.

Then guests will be those not part of the clique. Haven't decided if those that add them
to a conversation will route communication, or if user service will handle guest connections...

darcs-hash:20140620175558-c21fe-dd0b4d8dee790cb58859b3083f209805db2da75f