Major reorganization work towards supporting app execution on background threads.
Ignore-this: 5b696e002b84f1900452be3c5666ab7a

This commit is a major reorganization in the app interface.
I seperated the gui code from the backend app code. Now the backend talks via an front end interface,
and the frontend talks to backend via a backend interface. These interfaces do not assume
QT or Lua.

This is a step towards the goal of having the backend run on a seperate thread.

It also makes it possible in the future to use a different UI. The immediate goal
is to try QtQuick instead of QWidgets so that porting to phone becomes easier.

Also it opens up possibility of other languages besides Lua, though why would we want that? ;-)

darcs-hash:20141018062518-c21fe-9a81c276fd21b68b149b38eb2570717bf796ef79