fixed semantics in util::queue
Ignore-this: 197dc3f9f971ad8703e768ccdcc051ee

The semantics are that if the queue is done, the pop should return false.
The bug was that if there was something in the queue when the pop was called, it
returned true when the queue was done.

darcs-hash:20141026003304-c21fe-5b0f9d2d0709fcb463e880dcb8523ee245983580