Microphone is turned off by default.
Ignore-this: 9433a7f83a4f6b5052d3522514ccec89

When an app creates a microphone, the a mic button will pop up and the mic will be
disabled by default. The user will need to press the mic button to enable the mic
for that app.

Also if a script has an error, I added a widget to indicate that.

darcs-hash:bad4cc68e0ebfc0e6c731322aa23257182c3ff1d